Replace energy with corrected energy
Despite the corrections being in the miniAOD for 2017 data, ele.p4() by
default returns the uncorrected values.

This doesn't necessarily solve the smearing issue entirely, as the other
components of the 3 momentum are unaltered, but seemingly no userFloat
is provided for them. Waiting to hear back from the EGM contacts on this
issue.